4/9/25 Finished enemy pathfinding
finished implementing tutorial from https://www.youtube.com/watch?v=utLv2wKeCxs

Adjusted code so that NPC only progresses when a trailing joeBiden object get's within a certain distance of it so the movement is smoothed out.

Got the pathfinding to avoid walls

got the pathfinding to update every time it moves to a new node so that it trails the player more quickly.