4/16/25 Got path building to work
Now you build an entrance which then prompts you to build paths which then prompts you to build an exit.

All of the paths have to connect in order to be placed.

When a path is placed, it creates connections with the last path that was placed.

npc uses these connections to pathfind to exit.