4/22/25 Added child objects to arcade
added child objects onto arcade prefab that have nodes.

Each "arcade machine" in the arcade has a node connected to the door.

the door is connected to all the arcade machines and to the nearest pathway outside the door

getting a bug that won't allow npcs to make a path into the arcade now.