4/28/25 GOT JOB FILLING TO WORK
Finally got job filling to work.

npc follower creates a target position based off of the position of the next node the npc pathfinder is traveling to

This offset is randomized in a long range if it is traveling through the hallway and small if it's traveling inside a room.

The npc pathfinder generates a path to it's job using the nextNode instead of currentNode. Current Node was sending it back a full tile.