5/3/25 Buldozing pathways, money system
Added in the bulldozing of pathways.

Buildings will be automatically bulldozed when their pathway is removed

When a pathway is bulldozed, all pathways after it are as well

Exit is moved to the last pathway when a pathway is bulldozed.

Any npcs that are no longer on a pathway are killed.