5/26/25 Bug fixing, added roller rink
Fixed many a bug surrounding the targeting system mainly, which I ended up deleting in favor of a much simpler system.

Added a roller rink that is 4 tiles large and increased the speed of its customers forever.

Began adding a laser tag building that is 3 tiles large and has a waiting room.